py/requirements.txt
Brotli==1.0.9
certifi==2022.12.7
charset-normalizer==3.1.0
idna==3.4
mutagen==1.46.0
pycryptodomex==3.17
python-vlc==3.0.18121
requests==2.28.2
urllib3==1.26.15
websockets==10.4
yt-dlp==2023.3.4
